Add files via upload
With the rapid urbanization and development of big cities and towns, the graph of crimes is rising rapidly. This phenomenal rise in offences and crime in cities is a matter of great concern. There are robberies, murders, smuggling, rapes and crimes which one can't think of.

Digitization of Records: Crime reporting solutions used today are trivial. Crime reports are still stored in paper records and evidences such as media files are stored in CD's or DVD's. When the case progresses, the paper records and evidences tend to increase and makes it difficult to manage. Also, if a previous case is reopened after a long time, there is a hard time finding all the records regarding the case.

Anonymity: Many of the crimes that happen in India are not reported to the police. There can be several reasons why a victim do not report a crime like he don't want to get involved in a police case or he don't want to disclose his identity. Despite that, for every crime only the police is held responsible.

Reporting of Cyber Crimes: According to a report by Times of India, from the global ransomware attacks that hit hundreds of systems to phishing and scanning rackets, at least one cybercrime was reported every 10 minutes in India in the first six months of 2017. That's higher than a crime every 12 minutes in 2016. According to the Indian Computer Emergency Response Team (CERT-In), 27,482 cases of cybercrime were reported from January to June. With more Indians going online, cyber experts said putting in place critical infrastructure to predict and prevent cybercrimes was crucial. There is a need of dedicated cyber cell to address cyber crimes reported by the people. In our current system, there is no way to report a cyber crime online.

Crime Analytics and Statistics: Crime Statistics and Analytics provide a detailed and informative view of different aspects of crimes. They can be used to determine crime prone areas so that effective measures can be taken. We can also use them to monitor the change in crime rates so that we can implement effective strategies to counter the growth of crimes. Most Wanted Person’s Online Directory: The person with a criminal record could be dangerous to the society. There is a need for online directory listing of the criminals on the most wanted list of the police which will maintain a list of records of all the criminals with bounty reward.


Tech Stack
Python
Django
Javascript/Jquery
Postgresql
Bootstrap 4
Django Rest Framework
Bhamashah API
Google Maps API
Google Charts API
Other Javascript libraries/ packages

Instructions for executing locally:
cd Online-Crime-Reporting-System
pip3 install -r requirements.txt
cd src/
./manage.py makemigrations
./manage.py migrate
./manage.py runserver